Move common runtimes definitions into a separate crate
