+++ node/cli/Cargo.toml
@@ -20,6 +20,7 @@ service = { package = "gear-service", path = "../service", default-features = fa
 # Gear Runtimes
 gear-runtime = { path = "../../runtime/gear", optional = true }
 vara-runtime = { path = "../../runtime/vara", optional = true }
+gear-runtime-common = { path = "../../runtime/common", default-features = false }
 
 # Substrate client
 sc-cli = { version = "0.10.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", features = [
