+++ runtime/common/src/apis.rs
@@ -1,0 +1,224 @@
+// Copyright 2019-2022 PureStake Inc.
+// This file is part of Moonbeam.
+
+// Moonbeam is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Moonbeam is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
+
+#[macro_export]
+macro_rules! impl_runtime_apis_plus_common {
+	{$($custom:tt)*} => {
+		impl_runtime_apis! {
+			$($custom)*
+
+			impl sp_api::Core<Block> for Runtime {
+				fn version() -> RuntimeVersion {
+					VERSION
+				}
+
+				fn execute_block(block: Block) {
+					Executive::execute_block(block);
+				}
+
+				fn initialize_block(header: &<Block as BlockT>::Header) {
+					Executive::initialize_block(header)
+				}
+			}
+
+			impl sp_api::Metadata<Block> for Runtime {
+				fn metadata() -> OpaqueMetadata {
+					OpaqueMetadata::new(Runtime::metadata().into())
+				}
+			}
+
+			impl sp_block_builder::BlockBuilder<Block> for Runtime {
+				fn apply_extrinsic(extrinsic: <Block as BlockT>::Extrinsic) -> ApplyExtrinsicResult {
+					Executive::apply_extrinsic(extrinsic)
+				}
+
+				fn finalize_block() -> <Block as BlockT>::Header {
+					Executive::finalize_block()
+				}
+
+				fn inherent_extrinsics(data: sp_inherents::InherentData) -> Vec<<Block as BlockT>::Extrinsic> {
+					data.create_extrinsics()
+				}
+
+				fn check_inherents(
+					block: Block,
+					data: sp_inherents::InherentData,
+				) -> sp_inherents::CheckInherentsResult {
+					data.check_extrinsics(&block)
+				}
+			}
+
+			impl sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block> for Runtime {
+				fn validate_transaction(
+					source: TransactionSource,
+					tx: <Block as BlockT>::Extrinsic,
+					block_hash: <Block as BlockT>::Hash,
+				) -> TransactionValidity {
+					Executive::validate_transaction(source, tx, block_hash)
+				}
+			}
+
+			impl sp_offchain::OffchainWorkerApi<Block> for Runtime {
+				fn offchain_worker(header: &<Block as BlockT>::Header) {
+					Executive::offchain_worker(header)
+				}
+			}
+
+			impl sp_session::SessionKeys<Block> for Runtime {
+				fn generate_session_keys(seed: Option<Vec<u8>>) -> Vec<u8> {
+					SessionKeys::generate(seed)
+				}
+
+				fn decode_session_keys(
+					encoded: Vec<u8>,
+				) -> Option<Vec<(Vec<u8>, KeyTypeId)>> {
+					SessionKeys::decode_into_raw_public_keys(&encoded)
+				}
+			}
+
+			impl fg_primitives::GrandpaApi<Block> for Runtime {
+				fn grandpa_authorities() -> GrandpaAuthorityList {
+					Grandpa::grandpa_authorities()
+				}
+
+				fn current_set_id() -> fg_primitives::SetId {
+					Grandpa::current_set_id()
+				}
+
+				fn submit_report_equivocation_unsigned_extrinsic(
+					_equivocation_proof: fg_primitives::EquivocationProof<
+						<Block as BlockT>::Hash,
+						NumberFor<Block>,
+					>,
+					_key_owner_proof: fg_primitives::OpaqueKeyOwnershipProof,
+				) -> Option<()> {
+					None
+				}
+
+				fn generate_key_ownership_proof(
+					_set_id: fg_primitives::SetId,
+					_authority_id: GrandpaId,
+				) -> Option<fg_primitives::OpaqueKeyOwnershipProof> {
+					// NOTE: this is the only implementation possible since we've
+					// defined our key owner proof type as a bottom type (i.e. a type
+					// with no values).
+					None
+				}
+			}
+
+			impl frame_system_rpc_runtime_api::AccountNonceApi<Block, AccountId, Index> for Runtime {
+				fn account_nonce(account: AccountId) -> Index {
+					System::account_nonce(account)
+				}
+			}
+
+			impl pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance> for Runtime {
+				fn query_info(
+					uxt: <Block as BlockT>::Extrinsic,
+					len: u32,
+				) -> pallet_transaction_payment_rpc_runtime_api::RuntimeDispatchInfo<Balance> {
+					GearPayment::query_info(uxt, len)
+				}
+				fn query_fee_details(
+					uxt: <Block as BlockT>::Extrinsic,
+					len: u32,
+				) -> pallet_transaction_payment::FeeDetails<Balance> {
+					GearPayment::query_fee_details(uxt, len)
+				}
+			}
+
+			// Here we implement our custom runtime API.
+			impl pallet_gear_rpc_runtime_api::GearApi<Block> for Runtime {
+				fn calculate_gas_info(
+					account_id: H256,
+					kind: HandleKind,
+					payload: Vec<u8>,
+					value: u128,
+					allow_other_panics: bool,
+					initial_gas: Option<u64>,
+				) -> Result<pallet_gear::GasInfo, Vec<u8>> {
+					Gear::calculate_gas_info(account_id, kind, payload, value, allow_other_panics, initial_gas)
+				}
+			}
+
+			#[cfg(feature = "runtime-benchmarks")]
+			impl frame_benchmarking::Benchmark<Block> for Runtime {
+				fn benchmark_metadata(extra: bool) -> (
+					Vec<frame_benchmarking::BenchmarkList>,
+					Vec<frame_support::traits::StorageInfo>,
+				) {
+					use frame_benchmarking::{baseline, Benchmarking, BenchmarkList};
+					use frame_support::traits::StorageInfoTrait;
+					use frame_system_benchmarking::Pallet as SystemBench;
+					use baseline::Pallet as BaselineBench;
+
+					let mut list = Vec::<BenchmarkList>::new();
+					list_benchmarks!(list, extra);
+
+					let storage_info = AllPalletsWithSystem::storage_info();
+
+					(list, storage_info)
+				}
+
+				fn dispatch_benchmark(
+					config: frame_benchmarking::BenchmarkConfig
+				) -> Result<Vec<frame_benchmarking::BenchmarkBatch>, sp_runtime::RuntimeString> {
+					use frame_benchmarking::{baseline, Benchmarking, BenchmarkBatch, TrackedStorageKey};
+
+					use frame_system_benchmarking::Pallet as SystemBench;
+					use baseline::Pallet as BaselineBench;
+
+					impl frame_system_benchmarking::Config for Runtime {}
+					impl baseline::Config for Runtime {}
+
+					let whitelist: Vec<TrackedStorageKey> = vec![
+						// Block Number
+						hex_literal::hex!("26aa394eea5630e07c48ae0c9558cef702a5c1b19ab7a04f536c519aca4983ac").to_vec().into(),
+						// Total Issuance
+						hex_literal::hex!("c2261276cc9d1f8598ea4b6a74b15c2f57c875e4cff74148e4628f264b974c80").to_vec().into(),
+						// Execution Phase
+						hex_literal::hex!("26aa394eea5630e07c48ae0c9558cef7ff553b5a9862a516939d82b3d3d8661a").to_vec().into(),
+						// Event Count
+						hex_literal::hex!("26aa394eea5630e07c48ae0c9558cef70a98fdbe9ce6c55837576c60c7af3850").to_vec().into(),
+						// System Events
+						hex_literal::hex!("26aa394eea5630e07c48ae0c9558cef780d41e5e16056765bc8461851072c9d7").to_vec().into(),
+					];
+
+					let mut batches = Vec::<BenchmarkBatch>::new();
+					let params = (&config, &whitelist);
+					add_benchmarks!(params, batches);
+
+					Ok(batches)
+				}
+			}
+
+			#[cfg(feature = "try-runtime")]
+			impl frame_try_runtime::TryRuntime<Block> for Runtime {
+				fn on_runtime_upgrade() -> (Weight, Weight) {
+					// NOTE: intentional unwrap: we don't want to propagate the error backwards, and want to
+					// have a backtrace here. If any of the pre/post migration checks fail, we shall stop
+					// right here and right now.
+					let weight = Executive::try_runtime_upgrade().unwrap();
+					(weight, BlockWeights::get().max_block)
+				}
+
+				fn execute_block_no_check(block: Block) -> Weight {
+					Executive::execute_block_no_check(block)
+				}
+			}
+		}
+	}
+}
