+++ runtime/vara/Cargo.toml
@@ -55,6 +55,7 @@ frame-system-benchmarking = { version = "4.0.0-dev", default-features = false, g
 hex-literal = { version = "0.3.4", optional = true }
 
 # Internal deps
+runtime-common = { package = "gear-runtime-common", version = "0.1.0", default-features = false, path = "../common" }
 pallet-gear-scheduler = { version = "1.0.0", default-features = false, path = "../../pallets/gear-scheduler" }
 pallet-gear-messenger = { version = "1.0.0", default-features = false, path = "../../pallets/gear-messenger" }
 pallet-gear-program = { version = "2.0.0", default-features = false, path = "../../pallets/gear-program" }
@@ -98,6 +99,7 @@ std = [
 	"pallet-transaction-payment/std",
 	"pallet-transaction-payment-rpc-runtime-api/std",
 	"pallet-utility/std",
+	"runtime-common/std",
 	"scale-info/std",
 	"sp-api/std",
 	"sp-block-builder/std",
