+++ scripts/src/test.sh
@@ -130,5 +130,5 @@ run_fuzzer() {
 
   # Run fuzzer
   RUST_LOG="essential,pallet_gear=debug,gear_core_processor::executor=debug,economic_checks=debug,gwasm=debug" \
-  cargo fuzz run --release $FEATURES --sanitizer=none $TARGET
+  cargo fuzz run --release "$FEATURES" --sanitizer=none "$TARGET"
 }
